+++ src/main/java/com/buuz135/hotornot/HotOrNot.java
@@ -23,6 +23,7 @@ package com.buuz135.hotornot;
 
 import com.buuz135.hotornot.config.HotLists;
 import com.buuz135.hotornot.network.SyncClientLists;
+import com.buuz135.hotornot.util.HotOrNotTab;
 import net.minecraft.creativetab.CreativeTabs;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.Mod.EventHandler;
